:bug: fix: AppContainer and migrage 'studio' p
refix
