+++ src/ProBuilder/container/AppContainer.tsx
@@ -1,9 +1,13 @@
 import { GetAntdThemeConfig, ThemeAppearance, ThemeMode } from 'antd-style';
 import { ThemeConfig } from 'antd/es/config-provider/context';
-
 import { FC, PropsWithChildren } from 'react';
-import { OverrideAntdGlobalStyles } from '../antd/override';
-import { ThemeProvider, createStudioAntdTheme, getStudioStylish, getStudioToken } from '../theme';
+import {
+  ThemeProvider,
+  createStudioAntdTheme,
+  getStudioStylish,
+  getStudioToken,
+} from '../../theme';
+import { OverrideAntdGlobalStyles } from '../components/antd/override';
 
 /**
  * @title 应用容器属性
@@ -27,9 +31,6 @@ export interface AppContainerProps {
   theme?: ThemeConfig | GetAntdThemeConfig;
 }
 
-// 带改造，需要和 Studio 同步的
-const STUDIO_UI_PREFIX = 'studio';
-
 export const AppContainer: FC<PropsWithChildren<AppContainerProps>> = ({
   children,
   theme,
@@ -37,7 +38,6 @@ export const AppContainer: FC<PropsWithChildren<AppContainerProps>> = ({
   themeMode,
 }) => (
   <ThemeProvider
-    prefixCls={STUDIO_UI_PREFIX}
     appearance={appearance}
     themeMode={themeMode}
     // 以下都是自定义主题
