+++ tests/demo.test.tsx
@@ -1,4 +1,3 @@
-import { AppContainer } from '@ant-design/pro-editor';
 import { render } from '@testing-library/react';
 import * as fs from 'fs';
 import { glob } from 'glob';
@@ -37,7 +36,7 @@ dirs.forEach((dir) => {
 
         if (!demoName) return;
         // console.log(`测试组件${dir} DEMO:${demoName}`);
-        const wrapper = render(<Demo.default />, { wrapper: AppContainer });
+        const wrapper = render(<Demo.default />);
         act(() => {
           vi.runAllTimers();
         });
